Reorganize shortest path algorithms in new file
